requirements.txt
discord.py==2.3.2
transformers==4.41.1
torch>=2.0.0
deep-translator==1.11.4